pyproject.toml
[tool.poetry]
name = "flask_template"
version = "0.2.0"
description = "A template for Flask applications with SQLalchemy, Migrations and REST API"
authors = ["Fabian Bühler"]
license = "Unlicense"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
flask = { extras = ["dotenv"], version = "^2.0.0" }
Flask-JWT-Extended = "^4.0.0"
Flask-Cors = "^3.0.10"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.0.0"
flask-babel = "^2.0.0"
flask-smorest = "^0.31.0"
tomlkit = "^0.7.0"

[tool.poetry.dev-dependencies]
black = "^21.5b1"
flake8 = "^3.9.2"
pytest = "^6.2.4"
flake8-docstrings = "^1.6.0"
flake8-bugbear = "^21.4.3"
Sphinx = "^4.0.0"
recommonmark = "^0.7.0"
sphinxcontrib-redoc = "^1.6.0"
sphinx-click = "^3.0.0"

[tool.poetry.scripts]
#flask = 'flask.cli:main'

[tool.black]
target-version = ["py37"]
line-length = 90
include = '\.pyi?$'

[tool.sphinx]
copyright-year = 2021
enable-autodoc = true
enable-autosectionlabel = true
enable-markdown = true
enable-todo = true
# update gitignore if you want to edit docs/changelog.md manually!
include-changelog = true
# update gitignore if you want to edit docs/readme.md manually!
include-readme = true

[tool.sphinx.intersphinx-mapping]
python = ["https://docs.python.org/3"]

[tool.sphinx.recommonmark]
auto-toc-tree-section = "Contents"
enable-math = false
enable-inline-math = false
enable-eval-rst = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
